refactor: moved around warnings
