+++ cognite_toolkit/_cdf_tk/validation/warning/__init__.py
@@ -1,3 +1,4 @@
-from .fileread import DataSetMissingWarning, DataSetMissingWarningList, SnakeCaseWarning, TemplateVariableWarning
+from ._base import DataSetMissingWarningList
+from .fileread import DataSetMissingWarning, SnakeCaseWarning, TemplateVariableWarning
 
-__all__ = ["SnakeCaseWarning", "DataSetMissingWarning", "DataSetMissingWarningList", "TemplateVariableWarning"]
+__all__ = ["SnakeCaseWarning", "DataSetMissingWarning", "TemplateVariableWarning", "DataSetMissingWarningList"]
