+++ cognite_toolkit/_cdf_tk/validation/warning/fileread.py
@@ -1,16 +1,13 @@
 from __future__ import annotations
 
-import itertools
-from collections import UserList
-from collections.abc import Collection
 from dataclasses import dataclass
 from functools import total_ordering
 from pathlib import Path
-from typing import ClassVar, Generic, TypeVar
+from typing import ClassVar
 
 
 @dataclass(frozen=True)
-class LoadWarning:
+class FileReadWarning:
     _type: ClassVar[str]
     filepath: Path
     id_value: str
@@ -19,7 +16,7 @@ class LoadWarning:
 
 @total_ordering
 @dataclass(frozen=True)
-class SnakeCaseWarning(LoadWarning):
+class SnakeCaseWarning(FileReadWarning):
     actual: str
     expected: str
 
@@ -49,7 +46,7 @@ class SnakeCaseWarning(LoadWarning):
 
 @total_ordering
 @dataclass(frozen=True)
-class TemplateVariableWarning(LoadWarning):
+class TemplateVariableWarning(FileReadWarning):
     path: str
 
     def __lt__(self, other: object) -> bool:
@@ -68,7 +65,7 @@ class TemplateVariableWarning(LoadWarning):
 
 @total_ordering
 @dataclass(frozen=True)
-class DataSetMissingWarning(LoadWarning):
+class DataSetMissingWarning(FileReadWarning):
     resource_name: str
 
     def __lt__(self, other: object) -> bool:
@@ -89,48 +86,3 @@ class DataSetMissingWarning(LoadWarning):
             return f"{type(self).__name__}: It is recommended to use a data set if source or destination can be scoped with a data set. If not, ignore this warning."
         else:
             return f"{type(self).__name__}: It is recommended that you set dataSetExternalId for {self.resource_name}. This is missing in {self.filepath.name}. Did you forget to add it?"
-
-
-T_Warning = TypeVar("T_Warning", bound=LoadWarning)
-
-
-class Warnings(UserList, Generic[T_Warning]):
-    def __init__(self, collection: Collection[T_Warning] | None = None):
-        super().__init__(collection or [])
-
-
-class SnakeCaseWarningList(Warnings[SnakeCaseWarning]):
-    def __str__(self) -> str:
-        output = [""]
-        for (file, identifier, id_name), file_warnings in itertools.groupby(
-            sorted(self), key=lambda w: (w.filepath, w.id_value, w.id_name)
-        ):
-            output.append(f"    In File {str(file)!r}")
-            output.append(f"    In entry {id_name}={identifier!r}")
-            for warning in file_warnings:
-                output.append(f"{'    ' * 2}{warning!s}")
-
-        return "\n".join(output)
-
-
-class TemplateVariableWarningList(Warnings[TemplateVariableWarning]):
-    def __str__(self) -> str:
-        output = [""]
-        for path, module_warnings in itertools.groupby(sorted(self), key=lambda w: w.path):
-            if path:
-                output.append(f"    In Section {str(path)!r}")
-            for warning in module_warnings:
-                output.append(f"{'    ' * 2}{warning!s}")
-
-        return "\n".join(output)
-
-
-class DataSetMissingWarningList(Warnings[DataSetMissingWarning]):
-    def __str__(self) -> str:
-        output = [""]
-        for filepath, warnings in itertools.groupby(sorted(self), key=lambda w: w.filepath):
-            output.append(f"    In file {str(filepath)!r}")
-            for warning in warnings:
-                output.append(f"{'    ' * 2}{warning!s}")
-
-        return "\n".join(output)
